Changelog and progress for Week4
Initially, as the UserProfile model now applies the AbstractUser
model from the Django framework (you will need to makemigrations
and migrate using manager.py again), I mute the warning that mentions
overriding fields to be taken precautions. Also, the password that
users typed when Sign up and Log in is now hashed.

Secondly, to solve the cross-field problem, we will be injecting
core.header package from Django framework, and subsequently, you
will be required to run pip3 install -r requirement.txt.

Then, to be able to store the user profile in frontend so that the
number of times for requesting user information can be reduced. And
hence, Vuex has been added to the project for frontend. And due to
that reason, you would probably run npm install before running the
project. The changes are mainly in Toolbar.vue and Login.vue.

Last thing last, the Log out action is encoded in the frontend,
since all user information are stored in localStorage. Log out is
simply clear the 'user' stored in localStorage.

Signed-off-by: chrisly <chrizly07@gmail.com>