Few updates in main page
Initially, in toolbar, to emphasize the sign up button, only it will be
the primary button. Also, we have removed the border for log in button.

Secondly, we removed the label above the search bar indicating which
operations might be undertaken, as placeholders do the same trick. And
an guest icon has been added to select box.